Johny B. Good: killset done. One to go...

+++ src/dataflow.py
@@ -5,17 +5,30 @@ import parse_instr
 class Dataflow(object):
     def __init__(self, graph):
         self.graph = graph
+        self.set_ins_names()
         self.create_sets()
-                 
+    
+    def set_ins_names(self):
+        """
+        Creates names for instructions, so that they can easily be identified
+        during optimisation. Using indexes would cause problems.
+        """
+        for block in self.graph.blocks:
+            for i,instr in enumerate(block.instructions):
+                instr.id = block.name + "_" + str(i)
+    
+    
     def create_sets(self):
+        """
+        Calls all subroutines to create the different sets in the right order.
+        """
         self.create_gen()    
+        self.create_kill()
         
     def create_gen(self):
         """
         Determines which instruction are in the gen set of each block
         """
-        #self.graph.get_block("$L4").print_block()
-        #for block in [self.graph.get_block("$L4")]:
         for block in self.graph.blocks:
             # {"",[]}
             gen = {}
@@ -31,12 +44,14 @@ class Dataflow(object):
                                 killown[key] = gen[key]
                             if key in gen:
                                 del gen[key]
-                        gen[block.name + "_" + str(i)] = instr.gen
+                        gen[instr.id] = instr.gen
             block.genset = gen
             block.killown = killown   
             
     def create_kill(self):
-        block.killset = []
+        """
+        Determines which instruction are in the kill set of each block
+        """    
         for block in self.graph.blocks:
             self.get_reach(block)
         
@@ -45,41 +60,51 @@ class Dataflow(object):
             #Check all nodes that can be reached
             for targetname in block.reach:
                 target = self.graph.get_block(targetname)
-                kills = []
-                
-                #For all instruction in the target block
-                for ins in target.instructions:
-                    #Check if they write to the same registers as a instruction
-                    #in the origional block
-                    for g in ins.gen:
-                        kills += check_regs(block.genset, g)
-                        kills += check_regs(block.killown, g)
-                #Remove any duplicates in the list
-                #kills = remove_duplicates(kills)
-                
-                #For all instructions that are overwritten (killed) in the 
-                #origional block, find the corresponding registers and add the 
-                #killed instruction to the killset of the target block.
-                for kill in kills:
-                    if kill not in target.killset:
-                        if kill in block.genset:
-                            target.killset[kill] = block.genset[kill]
-                        elif kill in block.killset:
-                            target.killset[kill] = block.killown[kill]
-                #for g in target.gen: 
-                #    for reg in target.gen[g]:
-                #        kills = check_regs(block.gen, reg)
-                #for g in target.killown: 
-                #    for reg in target.killown[g]:
-                #        kills += check_regs(block.gen, reg)                           
-                #kills = self.remove_duplicates(kills)
-                #for kill in kills:
-                #    if kill not in target.killset:
-                #        if kill in block.gen:
-                #            target.killset[key] = block.gen[key]
-                #        elif kill in block.killown:
-                #            target.killset[key] = block.gen[key]                            
-                            
+                if target:
+                    kills = []
+                    
+                    #For all instruction in the target block
+                    for ins in target.instructions:
+                        if type(ins)==Instr:
+                            #Check if they write to the same registers as a 
+                            #instruction in the origional block
+                            for g in ins.gen:
+                                kills += self.check_regs(block.genset, g)
+                                kills += self.check_regs(block.killown, g)
+
+                    
+                    #For all instructions that are overwritten (killed) in the 
+                    #origional block, find the corresponding registers and add 
+                    #the killed instruction to the killset of the target block.
+                    for kill in kills:
+                        if kill not in target.killset:
+                            if kill in block.genset:
+                                target.killset[kill] = block.genset[kill]
+                            elif kill in block.killset:
+                                target.killset[kill] = block.killown[kill]
+    def create_inout(self):
+        pass
+    
+    def print_sets(self):
+        for block in self.graph.blocks:
+            print "------------------------------------------------------------"
+            block.print_block()
+            print "\nGenset:"
+            for i in block.genset:
+                print str(i) + ": " + str(block.genset[i])                        
+            print "\nKillownset:"
+            for i in block.killown:
+                print str(i) + ": " + str(block.killown[i])   
+            print "\nKillset:"
+            for i in block.killset:
+                print str(i) + ": " + str(block.killset[i])                
+            print "\nInset:"
+            for i in block.inset:
+                print str(i) + ": " + str(block.inset[i])   
+            print "\nOutset:"
+            for i in block.outset:
+                print str(i) + ": " + str(block.outset[i])                                        
+
 
     def remove_duplicates(self, l):
         """
@@ -150,6 +175,7 @@ def main():
     flat = parse_instr.parse(flat)
     c = CFG(flat)
     d = Dataflow(c)
+    d.print_sets()
     d.get_reach(c.get_block("$L7"))
